+++ rr/a.js
@@ -21,7 +21,7 @@ class A extends RR {
   }
 
   toTinydns () {
-    return `+${this.get('name')}:${this.get('address')}:${this.get('ttl')}:${this.get('timestamp')}:${this.get('location')}\n`
+    return `+${this.get('name')}:${this.get('address')}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
   }
 }