+++ rr/mx.js
@@ -32,7 +32,7 @@ class MX extends RR {
   }
 
   toTinydns () {
-    return `@${this.get('name')}::${this.get('address')}:${this.get('weight')}:${this.get('ttl')}:${this.get('timestamp')}:${this.get('location')}\n`
+    return `@${this.get('name')}::${this.get('address')}:${this.get('weight')}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
   }
 }