+++ rr/ns.js
@@ -15,7 +15,7 @@ class NS extends RR {
   }
 
   toTinydns () {
-    return `&${this.get('name')}::${this.get('address')}:${this.get('ttl')}:${this.get('timestamp')}:${this.get('location')}\n`
+    return `&${this.get('name')}::${this.get('address')}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
   }
 }